[audio] Convert all calls to GetChannel from C to LLCPP
The audio driver interface FIDL definition includes a method (GetVmo)
with an error (GetVmoError). Errors are implemented in FIDL as
unions and since now unions are all effectively xunions, FIDL
interfaces with errors are not supported by the C bindings.
This change does not convert all users of the audio driver
interface to use the full FIDL interface, but rather converts the
GetChannel call used in the legacy manually serialized interface from
the C bindings to the LLCPP bindings.  Without this change, even users
of the C bindings that would not call GetVmo would fail to compile. To
avoid this compilation error the GetVmo method was added to an allowed
list that "allow C bindings for libraries with xunions to be valid C
code" (see fxr/338112).

Test: Listening to audio on Astro connected to USB headphones.

Change-Id: Ia093de1cccbe27b74751a84caf84a55aa4cb9d1e