[minfs] Convert truncate test to zxtest
The patch converts the tests in utest/fs/test-truncate.cc
from unittest to zxtest.

The tests in fs/test/posix/ will contain tests for posix like
behavior of the filesystems. It will be used by minfs and by
memfs. All other posix related tests in utest/fs/ will
eventually move here.

The patch retains the old tests and file just so that we still track
the reported bug in case the new changes don't reproduce the bug.
There is some loss of information in unittest which necessitates
the conversion to zxtest. Since the intention is to reproduce
the bug, I have tried to retain the code from test-truncate.cc
in truncate_test.cc. There is a lot of room for cleanup,
including removal of dead flags and restructuring code for better
names.

Bug: 41601

Change-Id: Ief9420095affdeb76f8c181e769ad5a6d38d6ba6