[zircon][fidl][zx] Extract built-in zx library
This moves the magically injected `library zx` out of fidlc, and into
a normal source file in //zircon/syscalls/zx.fidl. This allows the
syscall definitions to share the definition with other fidl, and also to
have the rest of the syscall definitions live in library zx (rather than
the fake "library zz" that was internally mapped to zx by kazoo).

Rather than have every fidl library have to specify a public_deps on zx,
the GN templates are modified to automatically include that dependency,
so previous behaviour is maintained.

Bug: 7802
Bug: 42717
Change-Id: I1c89840369508dc1dc07f013e1c1eae2bfc2cb97