[pave_no_bind] Make netls dependency explicit
The "pave_no_bind" test has an implicit dependency on the netls tool.
This hasn't surfaced as a test failure until now because
"limited_netsvc_tests" has netls as a runtime dep and it's normally run
in the same task as "pave_no_bind", so netls happened to be available
when this test ran. But we're trying to split that task into two shards
for efficiency, which causes this test to fail.

Bug: 43705
Change-Id: I12c2d9b7ee80a506a7db393daf92d7a4aba81d58