[earcons] Make volume earcons correct volume
The earcons played on volume changes currently play at the volume that
it was changed from, not changed to. This adds a fix so that it plays at
the correct volume.

Fixed: 43075, 44197

Tests: Added fake for SoundPlayer and tests to make sure that the
soundplayer receives sounds on volume changes.

Empty sound files added as test sound files expected to be loaded by the
Player.

fx run-test setui_service_tests -- --test

Change-Id: I1585f02f3d3380be4a9804dfeae0932898895974