[cobalt] Remove unused CreateLogger factory method that takes a ProjectProfile.
All clients are using the CreateLoggerFromProjectId method, so the
CreateLogger method is unused and can be removed. Also the Cobalt 1.1
plans don't include support for creating a Logger from a client's
version of the cobalt registry, so it is better to remove this type
of Logger creation before it starts to be used by someone.

Tested: no new code
Change-Id: I71889f22c952cb5e6ea3df8d4e837627d7b6b48f