[netstack2][dhcp] Separate send and recv endpoints
Create dedicated endpoints for read and writes when sending DHCP packets
from the IPv4 unspecified source address so the endpoints can each bind
to different addresses - the write endpoint binds to the unspecified
address and the read endpoint binds to the broadcast address (which is
where replies will be sent in response to DHCP packets with the
unspecified source address). The write endpoint needs to explicitly bind
to the unspecified address because it is marked as NeverPrimaryEndpoint
in this change and will not be used unless explicitly bound to.

Test: fx run-test netstack_gotests -t netstack_dhcp_gotests -- \
      --test.run TestIPv4UnspecifiedAddressNotPrimaryDuringDHCP

Change-Id: I90a4f77eadfccd0e88b18368331f003e473b9258