[fidlcat] Allow custom flags in shell script
All flags passed after "run" are now passed as is to fidlcat, before
run. It is still possible to use extra -- to pass flags as is to
fidlcat.

(Note that extra -- are not compatible with run: both will stop the
processing by the script of the following flags, while both require
some processing by the shell.)

Change-Id: Icea4f1aaba78a58a2996e53804331e00b39b9b39