[camera][isp] ISP stream shutdown (part 2)
This is part 2 of the CLs to implement stream shutdown in the ISP.

This CL adds a Shutdown() method to the OutputStream. The camera
controller will use Shutdown() to notify the ISP that a stream is no
longer needed. All accesses to the stream's output buffers will stop, and
all resources will be cleaned up.

Upon the completion of this work, the ISP notifies the caller by
invoking the shutdown_callback.

Note: This CL adds the interfaces and ensures proper compilation.
Subsequent CLs will implement the required functionality.

Bug fxb/44126

Testing: Code compiles. No functional changes.

Change-Id: Ie114433e415ec579308292abf280c5a67fef6e78