[benchmarks] Add example test for testing regression detection and graphing
Add an example test whose running time we can deliberately change in
order to test regression detection.

My particular motivation here is to check that we can preserve the
history when we rename tests on the Catapult performance dashboard
(chromeperf).

Note that perfcompare indicates that this changes the performance
results of some other tests, which is odd.

Bug: 44174
Test: zircon_benchmarks is run on the CQ
 * local testing: "fx shell zircon_benchmarks -p --filter Loop"
Change-Id: I18bfbf210200bde95d0dad4b05ebf572eccfafba