[aml-rawnand] destroy interrupt thread in tests
The interrupt handler thread wasn't being destroyed in tests. Most of
the time this didn't matter, as the zx::interrupt destructor would kick
the thread out of its wait, but there was a window where the
zx::interrupt could get destroyed and then the thread would start trying
to wait on it, leading to use-after-free.

MULTIPLY: `[
  {
    "name": "aml-rawnand-unittest",
    "total_runs": 40
  }
]`

Bug: 44450
Change-Id: I2c7390ad3e8545d6c1b73e996c180f75543aac83