[ulib] Make unittests more static analysis friendly
This change reduces the number of false positives we see on unittests
when using the clang static analyzer. The idea is the following: when a
unittest fail the failure path should not be analyzed, it is a known
failure (after executing the tests) and every further warning on that
path is spurious.

Change-Id: Id15f2983ee1b3c9e3f0fc2c1c68f54e61661e0e0