[components][cml] Add directory `subdir` feature
This feature allows a directory capability (use, expose, or offer) to
optionally specify a `subdir`. The subdir is a path relative to the
directory root; when the directory capability is routed, the target
will receive only that subdir.

This feature also supports nesting. That is, you can route a directory
with a `subdir` that itself was routed with a `subdir`. The effect is
additive; i.e., the new subdir path will be appended to the old one.

The `subdir` feature is useful in cases where a component wishes to
reduce the scope of the directory it's routing, or as a convenient way
of injecting resources that reside in a subdirectory. For example, to
route config-data in the CFv2 topology, we can route the entire
config-data package directory, up to the components that use
config-data, at which point the parent can offer the appropriate subdir.

This change adds the plumbing for component manifests, but does not
actually implement the feature.

Change-Id: If523540d69d27edd344111103b70ccf14712e153