[scenic][root presenter] Fix for out of bounds presentation shutdown
When ReplacePresentationWith() is called, there is a race between the release of the root presenter's ViewHolder, and the release of the child Presentation's View. if the ViewHolder is released first, then the child will receive a kViewDisconnected message. If the child is released first, then we will receive a kViewDisconnected message. However, since we have already removed the Presentation from presentations_, this message will fail to find an associated Presentation object. This is not a bug, but the previous code would crash in this situation.

Fixed: 44364

Change-Id: If640b7b9d6a62b4735ebb69355331b2fee2b79ff