[debugger] Hook up breakpoint size.
Allow hardware breakpoints to have a size. Parse the size as an input
and display it when printing breakpoints and it applies.

Hooks up the breakpoint type strings as described in the documentation.
Previously this used old hardcoded strings that didn't match the
documentation.

Add a test for creating breakpoints with alternate type and size.

This does not automatically size watchpoints based on the expression.
Currently everything will default to 4 bytes as before. Auto sizing will
be added in a followup.

Use std::to_string in some touched code instead of fxl::StringPrintf.

Bug: 44196
Change-Id: I70481660421ce8eeb38e66d0b988221d52a490ec