[llcpp] Enforce LSB of pointers is 0
This is temporary - it will enforce that LSB is 0
during the migration to tracking_ptr.

tracking_ptr checks for this, but before everything
switches it is useful to enforce to make the migration
easier. Before all types are switched to tracking_ptr,
only unowned_ptr will be permitted (not the unique_ptr
case with LSB of 1) to avoid situations where the LSB
is set to 1 for a non-tracking_ptr.

Change-Id: Ief2e5f3d7142d0c9a73d0ebf0e2ef567cd152e0c