[settings] [tests] Use struct for audio fakes
Refactor the settings service audio tests to use a struct for storing
the fakes. That way when the fakes are created, the tests only have to
worry about the services that they need, and not all of them. It also
reduces the amount of code required to add a new fake.

Fixed: 44612
Tests: Refactored tests, all still pass
fx run-test setui_service_tests -- --test

Change-Id: I7143ccb1fb4ec4c2bc834d0cb8e687230b33e9c9