[PowerManager] Builder pattern for node creation
The current method of creating nodes doesn't scale well when there are
multiple optional arguments. The current approach uses constructor
variants to allow the optional arguments to be passed in (e.g., `new`,
`new_with_proxy`, etc.). The change is to use a builder pattern to
provide a cleaner and more scalable way to create nodes when optional
arguments are desired.

Test: fx run-test power_manager_tests
Change-Id: I7e74d31d0051ef92413c68e11ef94e0995a50a2b