[exception_broker] Load filters from startup.
Exception Broker can be given a config file that determines whether the
process limbo is active on startup. This patch augments that config json
file to also specify which filters to be active. Filters determine which
processes will be caught by limbo, so this lets the system specify those
from startup.

If no filter are set, the system will catch everything.

TEST=Add unit test.
     fx run-test exception_broker_tests

Change-Id: I139d74000737408c6218436d20569793721b69ad