[kazoo] Support excluding functions from analysis
There are some complex syscalls that are hard to model in the clang
static analyzer. While it is possible that we will improve the modeling
of those functions in the future, we should have a facility to supress
false positives from certain functions. An alternative would be to not
to annotate the problematic functions but that way we would loose all
the information we have from the annotation. This approach preserves the
information about ownership while also tells the analyzer to not to care
about these functions.

Change-Id: Ieddee44fb58ce7d53a8a5cf6ebcd7cc3add3657c