[escher] Remove Rounded Rect Factory
This CL completely removes the RoundedRectFactory from Escher/Scenic.
It is rebased on top of another CL which removes the RoundedRectFactory
dependency from the Snapshotter, which was its only client and was the
reason it had not been deleted thus far.

This results in a memory-usage performance boost for Scenic, as it no
longer allocates an unncessary mesh for each RoundedRect.

Testing: This is a pure refactor, no testing required.

SCN-1168: done
SCN-1429: done

Change-Id: I1ecf8b44886e23f543a0fdb25c1f051fe5d8b03d