[debugger] Plumb over R/W watchpoints.
A bit a more plumbing was needed to correctly activate this feature.
This also plumbs correctly some register state that wasn't being read
that enables the agent to correctly identify an exception as watchpoint
(before they were all read as HW breakpoints). With this change the
correct breakpoint is communicated up.

TEST=Manual

Change-Id: I6f7c7342e6e963f555d0212df90435f871fcabb0