[FDF]Revamp the resume driver hooks and fidl calls
This change does the following

1. resume was taking an input parameter of device
power state. We now have device performance states and resume hook will
be called with a requested performance state.

2. Resume hooks are made asynchronous. Resume hooks talk to device and
submit I/O as part of resuming to a particular performance state. Lets
not block the devhost thread while that happens.

3. The semantics of the resume are changed. If set_performance_state is
called on a device, the performance state is rembered and the device
will be resumed to this state from a suspend.

Change-Id: I638768fe5bb78c953629ceaf62a459153a3a1554