[fidl][go] generate WithCtx Stub type
Step 1 of transitioning from the current `DispatchImplWithCtx()` stub
method to the new `DispatchImpl()` stub method.

The existing generated Stub temporarily has both methods, with the new
method delegating to the existing one. This Stub will eventually be
removed.

We also introduce a new stub (suffixed `WithCtx`) which only has the
new `DispatchImpl()` stub method. It uses the new handler interface,
which passes down a Context to the application-level code. This new
stub makes it possible to subsequently transition app-level code.

Bug: 39618
Test: fx run-host-tests fidlgen_golang_test
Change-Id: I31a82098342f8610773c545462fb65d747b6e4ce