[zircon][debugger][arm64] Plumb in LSC for R/W watchpoints.
R/W watchpoints are set by changing the LSC bits.
A change in zircon was needed because it was always masking out those values.
This patch removes that masking and removes a check in the debugger
gating the feature.

NOTE: There is an apparent race with R/W watchpoints within the
debugger. Those can be handled in subsequent CL.

TEST=Manual

Change-Id: I5b94111286143628b5953ed953264f88aeaa86fd