[debugger] Debug Agent can "Kill" limbo processes.
Before, as limbo process handles do not have ZX_RIGHT_DESTROY,
attempting to kill the process would give a ZX_ERR_BAD_ACCESS.

With this change, the agent under hood detaches from the process and
waits it to re-enter limbo. At that time, it recognizes that it is one
of the process to kill and releases it from limbo, thus killing it. This
is transparent to the client.

TEST=Add unit test.
     fx run-test debug_agent_tests

Change-Id: I2f5786e91f389e90b99c83a7394cca91504f9261