[audio_core] Replace AudioLink with LinkMatrix
This change replaces AudioLink with LinkMatrix. AudioLink is integral to
AudioCore. It is how all AudioObjects (input and output devices and
renderers and capturers for API users) are linked together to
communicate samples.

AudioLink gives both ends of the link a reference to each other.
Validating and accessing links is facilitated with coupled logic in
AudioObject.

LinkMatrix, the replacement integrated in this CL, stores all reference
in one place, weakly. The Routegraph becomes the sole owner of API
AudioObjects. All logic regarding creating, modifying, and accessing
links lives in one well tested place.

Bug: 39426
Fixed: 43514

TEST: Tests run in asan
      E2E run
      Manual tests
      fx run-test audio_core_unittests
                  audio_fidelity_tests
		  audio_pipeline_tests
		  audio_device_tests
		  audio_loopback_tests
		  audio_fidl_tests
Change-Id: If7948ddd27a6c9f96ffff0c0baa3576857c620c2