[pm] Clarify how help works in pm
When running `pm -help` one might be confused by lack of documentation
on sub commands. It very much appears to be complete documentation for the tool
despite this not being the case. To get more specific help you need to run
`pm <subcommand> -help` instead. I've added a line to the useage string to
document this.

Bug: 43114
Change-Id: I58383914eaa41933aca3a78fcd6a4006990bf3b3