[earcons] Play volume-changed earcon on press
There is no earcon played when pressing the volume up button when volume
is already at max. This changes it so that it plays when volume is at
max.
Fixed: 44367
Tests: Added a test for setting the max volume multiple times and
checking the play count.
fx run-test setui_service_tests -- --test

Change-Id: I05381100c99cbdbbead94c24beab4f60739727f6