[kernel][arm64] Build with -mtp=el1 instead of -mcmodel=kernel
The -mtp=el1 switch is now available in Clang, and it has the
same effect that is why we added -mcmodel=kernel support for
the aarch64-fuchsia target.  Just use -mtp=el1 for the kernel,
as the thread pointer ABI should be independent of other code
model issues.  Since -mcmodel=kernel is otherwise identical
to -mcmodel=small and that's already the default, drop it.

Test: compiled kernel is identical before and after
Change-Id: Iab075fd7b8cd6a726363c6e4c1df7d6f09c24cbe