[nelson][audio] Add playback support
We reuse Astro's AMLogic support, however the SOC's register
set was changed compared to Astro's (beyond base addresses).
Hence we added versioning to the audio AMLogic library with a
baseline supporting SOCs S905D2G and T931G, and a new version
that supports the changes made for S905D3G.
We use the codec support for the Maxim 98373 chip without modifications,
defining its usage in the board file as a composite device.
We also use the HIFI_PLL clock as a composite device.
Also fixes unused MCLK setting for AMLogic audio and
fixes name for Astro input audio device.

Test: audio uapp tone playback on Nelson and Astro.
Test: runtests -t device-enumeration-test on Nelson and Astro.
Change-Id: I120e36ddfd5fda3ec8540d3f3ae8359a9d938bdd