[components] Implement `subdir` feature
This change implements the `subdir` feature, which allows a directory
capability to specify a subdir that it uses/exposes/offers, instead of
the full directory.

This will be used to route subdirs of `config-data`. See
03ced2ceb99a9387d167ba8ffc13192604001d66 for more information.

This change includes some refactors to routing:

- Represent the current capability state in a more structured way,
varied by capability type (CapabilityState). Previously, this was just a
`rights_state` field in WalkPosition.

- Fix a bug causing a crash when returning an error containing a
capability path for capabilities without paths.

- Restructure walk_expose_chain to stop special-casing the first expose
declaration. Instead, model the first expose declaration as a new routed
capability type, ComponentCapability::UseExposed. This makes it easier
to differentiate between the initial capability state and the state
transition that happens when walking the first expose.

- Represent paths internally as PathBuf rather than String. This is a more
structured representation, and makes it easier to build paths.

Change-Id: I5d9cece8fbce710b3598f92127237af22c8aaf61