[fidl][go] allow binding a WithCtx impl
Add a AddWithCtx() method on the Server type in the generated Go FIDL
code which takes an implementation of the WithCtx interface rather
than the old sans-context interface. This allows implementations of
WithCtx interfaces to be bound to channels.

This is the 3rd change in a series that will add a Context parameter
to all Go FIDL handlers.

Bug: 39618
Test: fx run-host-tests fidlgen_golang_test
Change-Id: Ib7efdee2b6786f7d7b62cfc270b6ce23b0f7fe7e