[libc] Add __sanitizer_memory_snapshot API
This requires a slight adjustment to the dynamic linker's
bookkeeping so it can identify RELRO regions precisely for exclusion
from reporting.  (Technically this only matters for the GNU linkers'
behavior wrt RELRO layout and not lld's, but it seems best to keep
the runtime semantics strictly to what the ELF headers say without
presuming the linker's policies for composing them.)

It also requires new bookkeeping to keep track of the list of all
threads (live or joinable).  Otherwise there's no safe way to look
at the thread data structures based on the thread pointer of a
suspended thread that could be in a startup race or an exit race, or
not be a libc-managed thread at all.

Some libc internals are cleaned up to be C++-compatible so new
libc code can be written in hermetic C++.

Bug: 43710
Test: new test cases in /boot/test/libc/sanitizer-test
Change-Id: I225a79a73cf87f718f9f0e265c1b52b15cd82b43