[debugger] Save inheritance paths.
Currently data in inherited objects is represented as an offset. But
virtual inheritance means that there isn't always a simple offset
representation.

Adds a new class to save a full inheritance path so these nonconstant
transitions can be represented. Add a new class hierarchy iteration
scheme that uses these objects. This is currently in parallel with the
old version and the new version is unused. It will be hooked up and
tested in a followup.

Bug: 41503

Change-Id: Icecf9ca0223e754c3bd762434ad95bef8d059fe1