[cobalt] Move the global metrics registry to a config_data rule.
This will be used to allow private configurations to be created on a
per-product basis.

This is a reland of 8855ae84cccc03ba8903945c2410278ab904bb35 which got
caught up in the revert of fxr/357887 even though it is unrelated.

Bug: 3767
Tested: no code changes, tested config_data build changes manually
Change-Id: I0a913b7ccae6ed01d9d152dcf5291219a62b4dca