[msd-vsl-gc] Change SubmitCommandBuffer to take MappedBatch.
This is so we can store the MappedBatch with the interrupt event.

This will also help with implementing
msd_execute_command_buffer_with_resources.

Since the tests don't create a context / connection, we use a
MockMappedBatch and pass in the address space / buffer to
SubmitCommandBuffer.

For the actual execute_command_buffer implementation,
we will retrieve these from the CommandBuffer to pass in.

BUG=43238

Change-Id: I1c9b864a76ff1e1cad0cd5187f4786c30887ee76