[loadbench] Replace "system-temp" with "isolated-temp"
system-temp is deprecated and going away very soon. As part of this
effort we are replacing all instances with "isolated-temp". Please see
the bug for more details. Your temp directory will still be accessible
inside the process namespace at /tmp however if you are acessing it from
the shell it will be located at:

1. /tmp/r/sys/component_name/ - For a component.
2. /tmp/r/sys/r/env_for_test/component_name/ - For a test.

Bug: 25081
Change-Id: If57bb5ab399f63e1267f4335308f952708d8e844