[debugger] Use InheritancePath in FoundMember.
Before the final offset of the data member in the collection was stored.
But virtual inheritance means that some members won't have constant
offsets.

Saves the full inheritance path in FoundMember and pass it in
VisitClassHierarchy. This allows the code to extract the full location.

These locations are not yet hooked up so the functionality should be
unchanged. This information will be used in a followup to resolve
virtual inheritance.

Bug: 41503

Change-Id: I3a4ad323814aff7a5cfca79fb7530c5cf5754515