[debugger] Add a to_string variant for hex strings.
We've started to use C++'s std::to_string more but it doesn't support
hex numbers which we have a lot of.

This adds a variant to_hex_string that creates a hex string with some
simple options for these uses.

Updates some callers that don't require other printf-style formatting
to use this new function.

Change-Id: I222985cae89f65dc48980e1617b026e44097124c