[zxcrypt] add Shred to f.h.block.encrypted protocol
This adds support to the encrypted block protocol to shred the volume,
in addition to unsealing and sealing it, and a corresponding
implementation in the zxcrypt driver (the only existing implementation
of the fuchsia.hardware.block.encrypted protocol).

This is intended to support usage by factory_reset and omaha_client.  In
particular, the latter will need to wipe user data without immediately
rebooting, so it can commit to the target bootloader/kernel/system image
after being sure that booting that image will not read data that it
cannot be expected to handle.

Bug: 43690
Change-Id: Ib2fb3d742bc711bf6c8eb0bb1133c33915298874