[escher] Static state setters in CommandBufferPipelineState.
Step 1 in a series of CLs to build Vulkan pipelines at startup, rather
than lazily as needed.

This CL gives CommandBufferPipelineState state-setting methods to match
those already in CommandBuffer, so that "eager" and "lazy" client code
looks as similar as possible.

Testing: add a test to verify that the corresponding methods produce
the same results.  This uncovered a bug where some Vulkan enum values
are too large to fit into the allotted bits; a DCHECK now prevents us
from stumbling on this in the future.

Change-Id: I0adfb7b5fad876c5a2945f0711be8ac95a8939df