[factory_reset] Fix test on Astro
On Astro, this test fails because the IsolatedDevmgr instance's fshost
attempts to get a zxcrypt key from the TEE.  This fails because the TEE
driver is not loaded in the IsolatedDevmgr, and it would be substantial
work to produce a simulated TEE.

This test shouldn't be relying on block-watcher to set up its zxcrypt
partition.  Instead, we can set up the zxcrypt volume ourselves, which
allows us to key it explicitly with a test key, rather than relying on
the ability to get the same system-wide key from the TEE.

Before this change, this test would fail when run on Astro.  It should
now pass on all supported hardware.

Bug: 43691
Change-Id: Iec6851bc47b4cff5cce1c1f4a7081d444d57fbe2