[fidlgen][go] Output fidl_handle_rights tag when field contains a handle
Handles can't directly hold this information in go, so it is being
placed on the parent field in table, struct or xunion. Since handles
can also appear in arrays and vectors, fields containing array or
vector of handle (recursively) also are annotated with handle_rights.

This information will be extracted for encoding and decoding.

Bug: 44473

Change-Id: Ib9c9bc42bbae49eea5728ad5a0669e9c5d2a9623