[run_test_component] Improve environment readability for test realms.
This change replaces the name env_for_test_### with test_env_ to reduce
huge path lengths which were not ergonomic when using the shell to copy
and paste between paths.

This change also prints the Realm Label: test_env_### out on test
executions to help developers trace back which temp directory this
execution refers to. In addition --static-realm-label= is renamed to
--realm-label= for convenience.

Change-Id: I154bd70cab35c047c53d3e8818ba3795ccc0f53e