[components] Fix flake caused by hub race
When an instance is destroyed, the hub listens to PreDestroyInstance to
remove that instance's directory from `children`. PreDestroyInstance is
triggered in the DeleteChild action, which may be executed after the
instance is removed from the list of live children. This can lead to a
race where the instance's directory is still in the hub at the time the
child is recreated, leading to a crash.

To fix, trigger PreDestroyInstance before marking the instance deleted.

Bug: 42966

Change-Id: I62f81ed030c163dd37a57ac6797e580e158dbcbc