[paver] Provide output on failure and success.
Currently, commands such as:

  install-disk-image init-partition-tables --block-device /dev/foo/bar

will produce no output, independent of whether /dev/foo/bar exists and
the operation succeeded, or if it doesn't exist and the operation
failed. While the binary will return a non-zero error code in the
failure case, this can be hard to observe as a user.

This CL adds error messages in the failure cases:

  $ install-disk-image init-partition-tables --block-device /dev/...
  disk-pave:[RealMain] Failed to initialize partition tables: ZX_ERR_PEER_CLOSED

and also a success message if an operation succeeded:

  $ install-disk-image init-partition-tables --block-device /dev/...
  disk-pave: init-partition-tables operation succeeded.

Logging only changes.

TESTED: paved atlas
Change-Id: I5727459e7d6c4da9e82da419b631123e63a50a49