[build] flatten "runtime" group
This definition is no longer complete, and was always a little oddly shaped,
sometimes the amber additon was desired, other times not, and in all use
cases, the group was overspecified in the consumer. It is also now somewhat
inadequate as a result of component manager presence, which is handled
specially by the build.

Change-Id: I9c69fe5aac0c433b96312a40593e44ed410536ed