[pkg-resolver] create client with and without keyid_hash_algorithms
Before this patch, the resolver assumed that all tuf keys should have
the `"keyid_hash_algorithms": ["sha256"]` specified. However this field
was only added to check compatibility with python-tuf, and python-tuf is
[considering] getting rid of them, if they can figure out how to do it
without breaking their current users.

So we'd like to migrate away from using them to avoid having to have these
fields around for all time. This is the first step is to allow us to
verify the initial TUF metadata with two variations of the root TUF keys,
one with, and one without the keyid_hash_algorithms specified. This is
safe (as in we won't double count a key) as long as the metadata doesn't
list the same key multiple times with different keyids.

Once everyone has migrated over to the new metadata that doesn't mention
`keyid_hash_algorithms`, we can get rid of the call to
`PublicKey::from_ed25519_with_keyid_hash_algorithms`.

[considering]: https://github.com/theupdateframework/tuf/issues/848

Fixed: 44490
Change-Id: Ib84ca4551b9d68f322039215ba40996608d6ca58