[debugger] Allow expressions in "disassemble".
Sometimes one wants to disassemble relative to some address or the
address is indirect. Supporting the "*" notation in the disassemble
command allows this to be expressed more simply.

Make "break" and "disassemble" as having one parameter only which
eliminates the need to quote expressions.

Add a new variant of EvalLocalInputLocation that extracts the necessary
information from a Frame.

Add a disassemble command unit test.

Fix a bug in the MockRemoteAPI exposed by the new test that marked all
short reads as invalid. Short reads should be valid as long as there is
any data.

Change-Id: Ibe5820273ce79071a718a3d8d52b5e7ae1ff2e70