[iosched] Enable deferred op completions
This change enables the handling of deferred op completions. When the
async completion callback is called to signal an op's work is done, the
completion is deferred to be done on a worker thread.

Tests in test/main.cc take advantage of these changes when performing
async operations. Modified ServeTest to use prime numbers of input to
avoid aliasing in dequeue vs defer increments.
Added AsyncCompletionRaceTest to simulate a race between the return of
Issue() and the asynchronous completion of the issued op, which is fixed
by this change.

Test: iosched passes
Change-Id: Ideb15746f7d3bd3851869677df0683c946f8781c