[modular][testathon] Removed FXL_CHECK that was sometimes flaky but without the check, the test is no longer flaky in basemgr_launcher_tests.
This check was to make sure there was a connection to basemgr in order to
shut it down. In the tests, the basemgr is in a different realm. We expressly
check for the path in the tests so removing the check removes the
false-negative case.

TESTED: fx run-test basemgr_launcher_tests --gtest_repeat=50

Bug: 42848

Change-Id: I75ad4ccc0c1977b8a6e3f1799596ec15320d20b3