[components] Refactoring and new test for BeforeStartInstance
This change refactors the BeforeStartInstance event payload to take a
custom RuntimeInfo struct, instead of an Arc<Mutex<Runtime>>. This way,
the Runtime struct does not need to be wrapped in a mutex, and the hook
can't modify the Runtime's state.

Also add a test that concurrent bind calls do not cause duplicate start.
This exercises an edge case that was not tested before (except
non-deterministically in some integration tests).

Change-Id: I3c4383f102b14aa789a132502e91cfc4d20f39a4