[FDF] Enable clients to synchronously unbind
At the moment fuchsia.device.controller has ScheduleUnbind
api to schedule unbind which does not let clients know about the status
of unbind. Clients end up watching for the device removal. At times, the
unbind might fail, and the clients will infinitely wait in this case.
Another challenge with ScheduleUnbind is that, it is called on the
device that needs to be removed. Once the device is removed, there
is no way to inform the client about the status of unbind.

This changeset adds UnbindChildren to fuchsia.device.controller and this
api will be called on the parent. This will make it easy to inform the clients
about the status of the unbinding their driver. Once this api returns,
clients can be assured that the children are removed. In the case where
there are mutliple children(uncommon), we can exten the api to unbind a
particular child in some way.

Test: fx run-test driver_host_tests. Could not add a failure case,
because unbind driver hook does not return status which warrants a
series of changes.

Bug:44589

Change-Id: Ibc954d38157b702d16499f92388de9f8bd118945