[audio_core] Use IntermediateBuffer in MixStage
Update MixStage to accept an output Stream that will be used to
source output buffers for mixed frames. By default we use the
IntermediateBuffer Stream implementation that has semantics that match
previous buffer that was internal to MixStage.

This is to enable some MixStages to push frames directly into a
RingBuffer instead of an IntermediateBuffer in order to enable tapping
the loopback point into different positions in the pipeline.

As a part of this change, we convert the domain of the TimelineFunction
from integral-frames to fractional-frames to simplify how this component
interacts with other Streams.

Test: fx run-test audio_core_unittests
Change-Id: I7887a1118504c2cf56d6737a1dde9b02b013bab3