[brcmfmac] Add a RAII Netbuf class
This commit adds a Netbuf class, which holds a memory buffer and its
associated completion callback in a RAII manner.  This allows us to pass
in ethernet_netbuf_t instances obtained from the DDK interface deeper
into the driver, potentially reducing the number of data copies that
have to be made.

This commit includes:

* EthernetNetbuf, which manages the lifetime of a ethernet_netbuf_t
  instance obtained from the DDK interface.
* AllocatedNetbuf, which manages a buffer backed by an owned char[], as
  an adapter for use cases that require a Netbuf not backed by an
  ethernet_netbuf_t.

Bug: 40719
Test: build and run WLAN connect on Astro
Change-Id: If2e644d55806beeafed580355daf02a485283481