[kernel][sched] Switch to the unified fair/deadline scheduler.
Change the default scheduler on all arches to the unified fair/deadline
scheduler. The fair scheduler behavior is unchanged from the previous
fair-only scheduler, the vast majority of threads in the system will
continue to behave identically. A small number of threads have been
pre-migrated to the deadline scheduler ahead of the switch and will
gain tighter timing and latency after the switch.

The pre-migrated threads are in in Scenic, Magma, and Astro display.

Test: All system and e2e tests. Several months of burn in on multiple
      devices. Extensive micro benchmark and trace analysis.
Change-Id: Ia9e9917062f42b93e15694301823533438e8384d