[kernel][wdt] Add a watchdog timer pdev.
++ Define a watchdog timer pdev interface.
++ Implement a generic version which should be able to work on most systems,
   provided that the interface to the HW watchdog is based on 32 bit MMIO accesses,
   and does not require any complicated sequencing.  Bootloaders may pass the
   location of the watchdog registers and initial state of the watchdog using a
   ZBI entry.
++ Add a kconsole extension to allow manual exercising of the WDT.

Also included in the generic driver definition is the ability to read a platform
specific reboot reason from hardware.  While this is not strictly speaking a
"watchdog" task, it gets to come along for the ride as a bonus.

Tested by setting up the watchdog hardware on Sherlock, and adding a fake ZBI
entry in the early boot of the device.  Then, manually confirmed that everything
was WAI and that force-hanging the system did result in a watchdog induced
reboot of the target.

Next steps are to add code to handle various special cases (like, attempting to
disable the watchdog during a panic if the device is configured for
halt-on-panic).

Change-Id: I1c1763c705242c18869ac399cd5e8aebb579411b