[audio_core] Update domain of RingBuffer TimelineFunction
Migrate from 'ref clock to ring position byte' to 'ref clock to
fractional frame' to be more consistent with the rest of the Stream
implementations. This was originally kept as a function of bytes to
simplify integration with DriverOutput and AudioDriver which both were
already constructing TimelineFunctions in that domain.

Test: fx run-test audio_core_unittests
      fx run-test audio_pipeline_tests
Change-Id: I39319cbb51671305ae8041df38c74349527721d4