[fxtest] Add --realm option to specify a realm to run component tests inside.
A common usage developers want when working with component tests is the
ability to download artifacts from their tests such as logs. This is
mainly a ergonomics feature where a developer wants a fixed path where
they can scp a file from, re-run a test, scp again and repeat. This flag
enables that.

By default /tmp data from tests is dumped at:
/tmp/r/sys/r/test_env_<random_number>/component_name.

With this feature on: fx test --realm=foo my_component
/tmp/r/sys/r/foo/component_name

Will be the path. This fixed path is much more useful if you want to
just grab a file after running a component test.

Bug: 25081
Change-Id: I0d8f85e4828221c8a0e0f7ea47cd2b8be21969fd