[gn-sdk]Fix CQ errors in generate.py
The generate.py script removes the existing output directory to
ensure, no files built previously will contaminate this build.
The command to remove the output directory returns a status code
of 1 if the directory doesn't exist which causes CQ to fail will
pass locally. This change adds a protective "if block" to the
"remove directory" command to ensure the output directory will
only be removed if it already exists.

An empty test_project has also been added to prevent CQ from
failing when attempting to copy the contents of test_project
which will be added in a followup CL.

Change-Id: I65d2105185c82850032af652b189ee844fb5b7bb